add new winner list and show count

+++ Sketchtacts/ShowEventViewController.swift
@@ -12,11 +12,11 @@ import CoreData
 
 class ShowEventViewController: UIViewController {
     
-    @IBOutlet weak var winnerLabel: UILabel!
-    @IBOutlet weak var tableView: UITableView!
+    @IBOutlet weak var contactsTableView: UITableView!
+    @IBOutlet weak var countItem: UIBarButtonItem!
+    
     var _event: Event!
     var _context: NSManagedObjectContext!
-    var _winner: Person!
     var fetchResultsController: NSFetchedResultsController<Person>!
     
     func setup(event: Event, context: NSManagedObjectContext) {
@@ -26,7 +26,7 @@ class ShowEventViewController: UIViewController {
     }
     
     @IBAction func onExport(_ sender: Any) {
-        let fileName = "\(_event.name!)_contacts.csv"
+        let fileName = "\(_event.name?.replacingOccurrences(of: " ", with: "") ?? "")_contacts.csv"
         let path = NSURL(fileURLWithPath: NSTemporaryDirectory()).appendingPathComponent(fileName)
         var csvText = "First Name,Last Name,Email,Company,Winner\n"
         for person in _event.people?.allObjects as! [Person] {
@@ -49,26 +49,14 @@ class ShowEventViewController: UIViewController {
     
     override func viewDidLoad() {
         let fetchRequest: NSFetchRequest<Person> = Person.fetchRequest()
-        let predicate = NSPredicate(format: "event == %@", _event)
-        fetchRequest.predicate = predicate
-        let sortDescriptors: [NSSortDescriptor] = [NSSortDescriptor(key: "firstName", ascending: true)]
-        fetchRequest.sortDescriptors = sortDescriptors
-        fetchResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: _context, sectionNameKeyPath: nil, cacheName: nil)
+        let entriesPredicate = NSPredicate(format: "event == %@", _event)
+        fetchRequest.predicate = entriesPredicate
+        let primarySortDescriptor = NSSortDescriptor(key: "winFlag", ascending: false)
+        let secondarySortDescriptor = NSSortDescriptor(key: "firstName", ascending: true)
+        fetchRequest.sortDescriptors = [primarySortDescriptor, secondarySortDescriptor]
+        fetchResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: _context, sectionNameKeyPath: "winFlag", cacheName: nil)
         fetchResultsController.delegate = self
         try! fetchResultsController.performFetch()
-        
-        if _event.hasWinner {
-            let fetchRequest: NSFetchRequest<Person> = Person.fetchRequest()
-            let predicate = NSPredicate(format: "event == %@ && winFlag == true", _event)
-            fetchRequest.predicate = predicate
-            let sortDescriptors: [NSSortDescriptor] = [NSSortDescriptor(key: "firstName", ascending: true)]
-            fetchRequest.sortDescriptors = sortDescriptors
-            let fetchResultsController = NSFetchedResultsController(fetchRequest: fetchRequest, managedObjectContext: _context, sectionNameKeyPath: nil, cacheName: nil)
-            fetchResultsController.delegate = self
-            try! fetchResultsController.performFetch()
-            _winner = fetchResultsController.fetchedObjects?.first
-            updateWinnerLabel()
-        }
     }
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -82,22 +70,12 @@ class ShowEventViewController: UIViewController {
             ctrl.setup(delegate: self, event: _event)
         }
     }
-    
-    func updateWinnerLabel() {
-        if _event.hasWinner && _winner != nil {
-            winnerLabel.text = "\(_winner.firstName ?? "") \(_winner.lastName ?? "") - \(_winner.company ?? "")"
-        }
-    }
 }
 
 extension ShowEventViewController: RunDrawingViewControllerDelegate {
     func onDone(_ viewController: RunDrawingViewController, wasCancelled: Bool, winner: Person?) {
         if !wasCancelled {
-            if _event.hasWinner {
-                
-            }
             _event.hasWinner = true
-            _winner = winner
             winner?.winFlag = true
             try! _context.save()
         }
@@ -106,12 +84,31 @@ extension ShowEventViewController: RunDrawingViewControllerDelegate {
 
 extension ShowEventViewController: NSFetchedResultsControllerDelegate {
     func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        tableView.reloadData()
-        updateWinnerLabel()
+        contactsTableView.reloadData()
     }
 }
 
 extension ShowEventViewController: UITableViewDelegate, UITableViewDataSource {
+    func numberOfSections(in tableView: UITableView) -> Int {
+        return (fetchResultsController.sections?.count)!
+    }
+    
+    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
+        if _event.hasWinner && section == 0 {
+            return "Winners"
+        } else {
+            return "Entries"
+        }
+    }
+    
+    func tableView(_ tableView: UITableView, titleForFooterInSection section: Int) -> String? {
+        let count = fetchResultsController.sections![section].numberOfObjects
+        if _event.hasWinner && section == 0 {
+            return "\(count) Winners"
+        } else {
+            return "\(count) Entries"
+        }
+    }
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return fetchResultsController.sections![section].numberOfObjects
@@ -128,7 +125,15 @@ extension ShowEventViewController: UITableViewDelegate, UITableViewDataSource {
         var actions: [UIContextualAction] = []
         let delete = UIContextualAction(style: .destructive, title: "Delete") { [unowned self] (action, view, done) in
             let person = self.fetchResultsController.object(at: indexPath)
-            self._context.delete(person)
+            if self._event.hasWinner && indexPath.section == 0 {
+                person.winFlag = false
+                if self.fetchResultsController.sections?[0].numberOfObjects == 1 {
+                    self._event.hasWinner = false
+                }
+            } else {
+                self._context.delete(person)
+            }
+            //TODO:  UPDATE TRY WITH TRY CATCH
             try! self._context.save()
             done(true)
         }
